♻️ フェーズ1: エラーハンドリングとローディング状態を分離
- useErrorHandlerフックを作成してエラー処理機能を独立化
- useLoadingStatesフックを作成してローディング状態管理を独立化
- books-providerから400行以上のコードを責任別に分割
- 単一責任の原則に従ったリファクタリングを実施
